requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
python-multipart==0.0.9
pydantic==2.*
python-pptx==0.6.23
markdown-it-py==3.*
tiktoken==0.7.*
httpx==0.27.*
